update nlpmodels default to ConcreteStructs version
