+++ variants/Project.toml
@@ -1,5 +1,4 @@
 [deps]
-ADNLPModels = "54578032-b7ea-4c30-94aa-7cbd1cce6c9a"
 AmplNLWriter = "7c4d4715-977e-5154-bfe0-e096adeac482"
 ConcreteStructs = "2569d6c7-a4a2-43d3-a901-331e8e4be471"
 CondaPkg = "992eb4ea-22a4-4c89-a5bb-47a3300528ab"
@@ -7,11 +6,9 @@ Ipopt = "b6b21f68-93f8-5de0-b562-5493be1d77c9"
 Ipopt_jll = "9cc047cb-c261-5740-88fc-0cf96f7bdcc7"
 JuMP = "4076af6c-e467-56ae-b986-b466b2749572"
 MathOptSymbolicAD = "309f4015-3481-4d63-a8f9-aeb13adfe8eb"
-NLPModelsIpopt = "f4238b75-b362-5c4c-b852-0801c9a21d71"
 Optimization = "7f7a1694-90dd-40f0-9382-eb1efda571ba"
 OptimizationMOI = "fd9f6733-72f4-499f-8506-86b2bdd0dea1"
 PowerModels = "c36e90e8-916a-50a6-bd94-075b64ef4655"
 PythonCall = "6099a3de-0909-46bc-b1f4-468b9a2dfc0d"
 ReverseDiff = "37e2e3b7-166d-5795-8a7a-e32c996b4267"
 SparseDiffTools = "47a9eef4-7e08-11e9-0b38-333d64bd3804"
-Symbolics = "0c5d862f-8b57-4792-8d23-62f2024744c7"
